add TestSetup function to reset test user's Firebase collection move crucial CI/CD methods to TestSetup class
